+++ database/session.go
@@ -7,7 +7,7 @@ import (
 
 type Session struct {
 	ID         string    `json:"id"`
-	UserID     int       `json:"userID"`
+	UserID     int64     `json:"userID"`
 	ExpiryDate time.Time `json:"expiryDate"`
 }
 
@@ -47,7 +47,7 @@ func (conn *MongoConnection) CheckSession(id string) (*Session, bool) {
 	return session, true
 }
 
-func (conn *MongoConnection) GetSessionByUserID(userID int) (*Session, error) {
+func (conn *MongoConnection) GetSessionByUserID(userID int64) (*Session, error) {
 	for _, session := range Sessions {
 		if session.UserID == userID {
 			return &session, nil
